Refactor: single canonical class, proxy cpp files eliminate duplication
VescUart polling moved into UartBrakes.ino; DoLoop now accepts
(rpm, current, voltage, newData) from the sketch layer. Both sketch
folders hold a one-line proxy VESCSmarterBrakes.cpp that #includes
the root implementation, so there is one source of truth for the
control loop. CANBrakes proxy adds #define USE_SOFTPWM before the
include, switching SetDimmerPower to SoftPWMSet for pins 0/1.
