+++ UartBrakes/VESCSmarterBrakes.cpp
@@ -1,0 +1,3 @@
+// Proxy — Arduino IDE only compiles .cpp files within the sketch folder.
+// This pulls in the canonical implementation from the project root.
+#include "../VESCSmarterBrakes.cpp"
